Migrate enzyme Phone tests to React Testing Library
